move getTimeToStart to common utils
