+++ apps/client/src/views/timeline/timeline.utils.ts
@@ -198,9 +198,3 @@ export function getUpcomingEvents(events: OntimeRundown, selectedId: MaybeString
   };
 }
 
-/**
- * Utility function calculates time to start
- */
-export function getTimeToStart(now: number, start: number, delay: number, offset: number): number {
-  return start + delay - now - offset;
-}
